requirements.txt
argh==0.26.2
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-reversion==2.0.13
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
future==0.16.0
httplib2==0.9.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mixin==1.1
mysqlclient==1.3.13
pathtools==0.1.2
Pillow==5.2.0
pkg-resources==0.0.0
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.12
reload==0.9
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
watchdog==0.8.3
XlsxWriter==1.0.5
xlwt==1.3.0
